+++ psl-core/src/test/java/org/linqs/psl/reasoner/admm/term/LinearConstraintTermTest.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of the PSL software.
  * Copyright 2011-2015 University of Maryland
- * Copyright 2013-2017 The Regents of the University of California
+ * Copyright 2013-2018 The Regents of the University of California
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -46,13 +46,13 @@ public class LinearConstraintTermTest {
 		 * 
 		 * Constraint inactive at solution
 		 */
-		double[] z = {0.2, 0.5};
-		double[] y = {0.0, 0.0};
-		double[] coeffs = {1.0, 1.0};
-		double constant = 1.0;
+		float[] z = {0.2f, 0.5f};
+		float[] y = {0.0f, 0.0f};
+		float[] coeffs = {1.0f, 1.0f};
+		float constant = 1.0f;
 		FunctionComparator comparator = FunctionComparator.SmallerThan;
-		double stepSize = 1.0;
-		double[] expected = {0.2, 0.5};
+		float stepSize = 1.0f;
+		float[] expected = {0.2f, 0.5f};
 		testProblem(z, y, coeffs, constant, comparator, stepSize, expected);
 		
 		/*
@@ -60,13 +60,13 @@ public class LinearConstraintTermTest {
 		 * 
 		 * Constraint active at solution
 		 */
-		z = new double[] {0.7, 0.5};
-		y = new double[] {0.0, 0.0};
-		coeffs = new double[] {1.0, 1.0};
-		constant = 1.0;
+		z = new float[] {0.7f, 0.5f};
+		y = new float[] {0.0f, 0.0f};
+		coeffs = new float[] {1.0f, 1.0f};
+		constant = 1.0f;
 		comparator = FunctionComparator.SmallerThan;
-		stepSize = 1.0;
-		expected = new double[] {0.6, 0.4};
+		stepSize = 1.0f;
+		expected = new float[] {0.6f, 0.4f};
 		testProblem(z, y, coeffs, constant, comparator, stepSize, expected);
 		
 		/*
@@ -74,26 +74,26 @@ public class LinearConstraintTermTest {
 		 * 
 		 * Equality constraint
 		 */
-		z = new double[] {0.7, 0.5};
-		y = new double[] {0.0, 0.0};
-		coeffs = new double[] {1.0, -1.0};
-		constant = 0.0;
+		z = new float[] {0.7f, 0.5f};
+		y = new float[] {0.0f, 0.0f};
+		coeffs = new float[] {1.0f, -1.0f};
+		constant = 0.0f;
 		comparator = FunctionComparator.Equality;
-		stepSize = 1.0;
-		expected = new double[] {0.6, 0.6};
+		stepSize = 1.0f;
+		expected = new float[] {0.6f, 0.6f};
 		testProblem(z, y, coeffs, constant, comparator, stepSize, expected);
 	}
 	
-	private void testProblem(double[] z, double[] y, double[] coeffs, double constant,
-			FunctionComparator comparator, final double stepSize, double[] expected) {
+	private void testProblem(float[] z, float[] y, float[] coeffs, float constant,
+			FunctionComparator comparator, final float stepSize, float[] expected) {
 		List<LocalVariable> variables = new ArrayList<LocalVariable>(z.length);
-		List<Double> coeffsList = new ArrayList<Double>(z.length);
+		List<Float> coeffsList = new ArrayList<Float>(z.length);
 
 		for (int i = 0; i < z.length; i++) {
 			variables.add(new LocalVariable(i, z[i]));
 			variables.get(i).setLagrange(y[i]);
 
-			coeffsList.add(new Double(coeffs[i]));
+			coeffsList.add(new Float(coeffs[i]));
 		}
 		
 		LinearConstraintTerm term = new LinearConstraintTerm(variables, coeffsList, constant, comparator);
